upgrade to 10.0 with refactoring
