+++ web_pdf_preview/controllers/__init__.py
@@ -1,45 +1,1 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#    Web PDF Report Preview & Print
-#    Copyright 2012 wangbuke <wangbuke@gmail.com>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-from openerp.addons.web.controllers.main import Reports, serialize_exception
-from openerp.addons.report.controllers.main import ReportController
-import urllib
-from openerp.addons.web.http import Controller, route, request
-
-class WebPdfReports(ReportController):
-
-    @route(['/report/preview'], type='http', auth="user")
-    def report_download(self, data, token):
-        result = super(WebPdfReports, self).report_download(data, token)
-        result.headers['Content-Disposition'] = result.headers['Content-Disposition'].replace('attachment', 'inline')
-        return result
-
-
-class PreviewReports(Reports):
-
-    @route('/web/report', type='http', auth="user")
-    @serialize_exception
-    def index(self, action, token):
-        result = super(PreviewReports, self).index(action, token)
-        result.headers['Content-Disposition'] = result.headers['Content-Disposition'].replace('attachment', 'inline')
-        return result
-
-
-# vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
+from . import main
