+++ web_pdf_preview/views/web_pdf_preview.xml
@@ -1,12 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!-- vim:fdn=3:
--->
-<openerp>
-    <data>
-        <template id="assets_backend" name="web_pdf_preview assets" inherit_id="web.assets_backend">
-            <xpath expr="." position="inside">
-                <script type="text/javascript" src="/web_pdf_preview/static/src/js/web_pdf_preview.js"></script>
-            </xpath>
-        </template>
-    </data>
-</openerp>
+<odoo>
+    <template id="assets_backend" name="web_pdf_preview assets" inherit_id="web.assets_backend">
+        <xpath expr="." position="inside">
+            <script type="text/javascript" src="/web_pdf_preview/static/src/js/web_pdf_preview.js"></script>
+        </xpath>
+    </template>
+</odoo>
